Move unit test files
