+++ src/net/sf/memoranda/tests/ContactTest.java
@@ -1,4 +1,4 @@
-package test;
+package net.sf.memoranda.tests;
 
 import static org.junit.Assert.*;
